requirements.txt
about-time==4.2.1
accelerate==1.12.0
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
aiosqlite==0.22.1
alive-progress==3.3.0
annotated-doc==0.0.4
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.12.1
asttokens==3.0.1
attrs==25.4.0
audioread==3.1.0
autograd==1.8.0
av==16.1.0
banks==2.3.0
beautifulsoup4==4.14.3
brotli==1.2.0
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
cma==4.4.2
colorama==0.4.6
colorlog==6.10.1
contourpy==1.3.3
ctranslate2==4.7.1
cycler==0.12.1
dataclasses-json==0.6.7
datasets==3.6.0
decorator==5.2.1
defusedxml==0.7.1
Deprecated==1.2.18
deprecation==2.1.0
diffusers==0.37.0.dev0
dill==0.3.8
dirtyjson==1.0.8
distro==1.9.0
docling==2.76.0
docling-core==2.67.1
docling-ibm-models==3.11.0
docling-parse==5.5.0
docopt==0.6.2
easyocr==1.7.2
et_xmlfile==2.0.0
executing==2.2.1
Faker==40.8.0
fastapi==0.128.0
faster-whisper==1.2.1
ffmpy==1.0.0
filelock==3.20.0
filetype==1.2.0
flatbuffers==25.12.19
fonttools==4.61.1
frozenlist==1.8.0
fsspec==2025.3.0
ftfy==6.3.1
gradio==6.4.0
gradio_client==2.0.3
graphemeu==0.7.2
greenlet==3.3.1
griffe==1.15.0
groovy==0.1.2
grpcio==1.76.0
h11==0.16.0
h2==4.3.0
hf-xet==1.3.2
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.3
huggingface_hub==0.36.2
hyperframe==6.1.0
idna==3.11
ImageIO==2.37.2
imageio-ffmpeg==0.6.0
importlib_metadata==8.7.1
inquirerpy==0.3.4
ipython==9.11.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
Jinja2==3.1.6
jiter==0.12.0
joblib==1.5.3
jsonlines==4.0.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonref==1.1.0
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
kiwisolver==1.4.10rc0
lance-namespace==0.5.2
lance-namespace-urllib3-client==0.5.2
lancedb==0.29.2
langchain==1.2.7
langchain-classic==1.0.1
langchain-community==0.4.1
langchain-core==1.2.7
langchain-qdrant==1.1.0
langchain-text-splitters==1.1.0
langgraph==1.0.7
langgraph-checkpoint==4.0.0
langgraph-prebuilt==1.0.7
langgraph-sdk==0.3.3
langsmith==0.6.5
latex2mathml==3.78.1
lazy-loader==0.5
librosa==0.11.0
llama-cloud==0.1.35
llama-cloud-services==0.6.54
llama-index==0.14.13
llama-index-cli==0.5.3
llama-index-core==0.14.13
llama-index-embeddings-huggingface==0.6.1
llama-index-embeddings-openai==0.5.1
llama-index-indices-managed-llama-cloud==0.9.4
llama-index-instrumentation==0.4.2
llama-index-llms-openai==0.6.15
llama-index-readers-file==0.5.6
llama-index-readers-llama-parse==0.5.1
llama-index-vector-stores-lancedb==0.4.4
llama-index-workflows==2.13.1
llama-parse==0.6.54
llvmlite==0.46.0
lxml==6.0.2
markdown-it-py==4.0.0
marko==2.2.2
MarkupSafe==2.1.5
marshmallow==3.26.2
matplotlib==3.10.8
matplotlib-inline==0.2.1
mdurl==0.1.2
ml_dtypes==0.5.4
moocore==0.2.0
more-itertools==10.8.0
moviepy==2.2.1
mpire==2.10.2
mpmath==1.3.0
msgpack==1.1.2
multidict==6.7.1
multiprocess==0.70.16
mypy_extensions==1.1.0
natsort==8.4.0
nest-asyncio==1.6.0
networkx==3.4.2
ninja==1.13.0
nltk==3.9.2
nncf==2.19.0
num2words==0.5.14
numba==0.64.0
numpy==2.2.6
omegaconf==2.3.0
onnx==1.20.1
onnx-ir==0.2.0
onnxruntime==1.24.2
onnxscript==0.6.2
openai==2.15.0
openai-whisper==20250625
opencv-python==4.13.0.90
opencv-python-headless==4.13.0.92
openpyxl==3.1.5
openvino==2025.4.1
openvino-genai==2025.4.1.0
openvino-telemetry==2025.2.0
openvino-tokenizers==2025.4.1.0
optimum==2.1.0.dev0
optimum-intel==1.27.0.dev0+42d3f63
optimum-onnx==0.1.0.dev0
orjson==3.11.5
ormsgpack==1.12.2
packaging==25.0
pandas==2.3.3
parso==0.8.6
peft==0.18.1
pfzy==0.3.4
pillow==11.3.0
pip==26.0.1
platformdirs==4.5.1
pluggy==1.6.0
polyfactory==3.3.0
pooch==1.9.0
portalocker==3.2.0
proglog==0.1.12
prompt_toolkit==3.0.52
propcache==0.4.1
protobuf==6.33.4
psutil==7.2.1
pure_eval==0.2.3
pyarrow==23.0.1
pyclipper==1.4.0
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
pydantic-settings==2.12.0
pydot==3.0.4
pydub==0.25.1
Pygments==2.19.2
pylance==2.0.1
pylatexenc==2.10
pymoo==0.6.1.6
PyMuPDF==1.26.7
pyparsing==3.3.2
pypdf==6.6.2
pypdfium2==5.5.0
python-bidi==0.6.7
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.2.1
python-multipart==0.0.22
python-pptx==1.0.2
pytz==2025.2
pyvips==3.1.1
pywin32==311
PyYAML==6.0.3
qdrant-client==1.16.2
qwen-vl-utils==0.0.14
rapidocr==3.7.0
referencing==0.37.0
regex==2026.1.15
requests==2.32.5
requests-toolbelt==1.0.0
rich==14.3.1
rpds-py==0.30.0
rtree==1.4.1
safehttpx==0.1.7
safetensors==0.7.0
scikit-image==0.26.0
scikit-learn==1.8.0
scipy==1.17.0
semantic-version==2.10.0
semchunk==2.2.2
sentence-transformers==5.2.1
sentencepiece==0.2.1
setuptools==80.9.0
shapely==2.1.2
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
soundfile==0.13.1
soupsieve==2.8.3
soxr==1.0.0
SQLAlchemy==2.0.46
stack-data==0.6.3
starlette==0.50.0
striprtf==0.0.26
sympy==1.14.0
tabulate==0.9.0
tantivy==0.25.1
tenacity==9.1.2
threadpoolctl==3.6.0
tifffile==2026.3.3
tiktoken==0.12.0
tokenizers==0.21.4
tomlkit==0.13.3
tools==1.0.21
torch==2.8.0+cpu
torchao==0.16.0
torchaudio==2.10.0+cpu
torchvision==0.23.0+cpu
tqdm==4.67.1
traitlets==5.14.3
transformers==4.53.3
tree-sitter==0.25.2
tree-sitter-c==0.24.1
tree-sitter-javascript==0.25.0
tree-sitter-python==0.25.0
tree-sitter-typescript==0.23.2
typer==0.21.1
typing_extensions==4.15.0
typing-inspect==0.9.0
typing-inspection==0.4.2
tzdata==2025.3
urllib3==2.6.3
uuid_utils==0.14.0
uvicorn==0.40.0
wcwidth==0.5.0
Werkzeug==3.1.6
wheel==0.45.1
wrapt==1.17.3
xlsxwriter==3.2.9
xxhash==3.6.0
yarl==1.22.0
yt-dlp==2026.2.21
zipp==3.23.0
zstandard==0.25.0
